feat: added server:start command
removed app subfolder from configuration
updated dashboard ui
updated commands

+++ routes/admin.php
@@ -17,16 +17,20 @@ use App\Http\Controllers\Admin\SettingsController;
 use App\Http\Controllers\Admin\DashboardController;
 use App\Http\Controllers\Admin\ActivitiesController;
 use App\Http\Controllers\Admin\NotificationsController;
+use App\Http\Controllers\Admin\WalletController;
 
 /**
  * Admin routes
  */
 
 Route::groupPrefix('admin', function () {
-    Route::groupMiddlewares(['remember', 'auth', 'dashboard'], function () {
+    Route::groupMiddlewares(['remember', 'auth'], function () {
         Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard.index');
 
         Route::get('resources/users', [UsersController::class, 'index'])->name('users.index');
+        Route::get('resources/users/new', [UsersController::class, 'new'])->name('users.new')->protected(Roles::ROLE[2]);
+        Route::get('resources/users/{id}/edit', [UsersController::class, 'edit'])->name('users.edit')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
+        Route::get('resources/users/{id}/read', [UsersController::class, 'read'])->name('users.read')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
 
         Route::get('resources/medias', [MediasController::class, 'index'])->name('medias.index');
         Route::get('resources/medias/search', [MediasController::class, 'search'])->name('medias.search');
@@ -36,7 +40,7 @@ Route::groupPrefix('admin', function () {
         Route::get('resources/medias/{id}/read', [MediasController::class, 'read'])->name('medias.read')->where(['id' => 'num']);
         Route::get('resources/medias/{id}/download', [MediasController::class, 'download'])->name('medias.download')->where(['id' => 'num']);
 
-        Route::get('resources/tickets/new', [TicketsController::class, 'new'])->name('tickets.new')->protected(Roles::ROLE[1], Roles::ROLE[2]);
+        Route::get('resources/tickets/new', [TicketsController::class, 'new'])->name('tickets.new')->protected(Roles::ROLE[0]);
         Route::get('resources/?{user_id}?/tickets', [TicketsController::class, 'index'])->name('tickets.index')->where(['user_id' => 'num']);
         Route::get('resources/tickets/{id}/read', [TicketsController::class, 'read'])->name('tickets.read')->where(['id' => 'num']);
 
@@ -44,28 +48,28 @@ Route::groupPrefix('admin', function () {
         Route::get('account/notifications', [NotificationsController::class, 'index'])->name('notifications.index');
         Route::get('account/{id}/settings', [SettingsController::class, 'index'])->name('settings.index')->where(['id' => 'num']);
         Route::get('account/activities', [ActivitiesController::class, 'index'])->name('activities.index');
-
-        Route::groupProtected([Roles::ROLE[0], Roles::ROLE[1]], function () {
-            Route::get('resources/users/new', [UsersController::class, 'new'])->name('users.new');
-            Route::get('resources/users/{id}/edit', [UsersController::class, 'edit'])->name('users.edit')->where(['id' => 'num']);
-            Route::get('resources/users/{id}/read', [UsersController::class, 'read'])->name('users.read')->where(['id' => 'num']);
-
-            Route::get('resources/invoices', [InvoicesController::class, 'index'])->name('invoices.index');
-            Route::get('resources/invoices/new', [InvoicesController::class, 'new'])->name('invoices.new');
-            Route::get('resources/invoices/{id}/edit', [InvoicesController::class, 'edit'])->name('invoices.edit')->where(['id' => 'num']);
-            Route::get('resources/invoices/{id}/read', [InvoicesController::class, 'read'])->name('invoices.read')->where(['id' => 'num']);
-        });
+        
+        Route::get('resources/invoices', [InvoicesController::class, 'index'])->name('invoices.index');
+        Route::get('resources/invoices/new', [InvoicesController::class, 'new'])->name('invoices.new')->protected(Roles::ROLE[2]);
+        Route::get('resources/invoices/{id}/edit', [InvoicesController::class, 'edit'])->name('invoices.edit')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
+        Route::get('resources/invoices/{id}/read', [InvoicesController::class, 'read'])->name('invoices.read')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
+    
+        Route::get('account/wallet', [WalletController::class, 'index'])->name('wallet.index')->protected(Roles::ROLE[1], Roles::ROLE[2]);
     });
 
-    Route::groupMiddlewares(['remember', 'auth', 'dashboard', 'csrf', 'sanitize'], function () {
+    Route::groupMiddlewares(['remember', 'auth', 'csrf', 'sanitize'], function () {
+        Route::delete('resources/users/?{id}?/delete', [UsersController::class, 'delete'])->name('users.delete')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
+        Route::patch('resources/users/?{id}?/update', [UsersController::class, 'update'])->name('users.update')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
+        Route::post('resources/users/create', [UsersController::class, 'create'])->name('users.create')->protected(Roles::ROLE[2]);
+        Route::post('resources/users/export', [UsersController::class, 'export'])->name('users.export')->protected(Roles::ROLE[2]);
+
         Route::delete('resources/medias/?{id}?/delete', [MediasController::class, 'delete'])->name('medias.delete')->where(['id' => 'num']);
         Route::patch('resources/medias/?{id}?/update', [MediasController::class, 'update'])->name('medias.update')->where(['id' => 'num']);
         Route::post('resources/medias/create', [MediasController::class, 'create'])->name('medias.create');
 
         Route::delete('resources/tickets/?{id}?/delete', [TicketsController::class, 'delete'])->name('tickets.delete')->where(['id' => 'num']);
         Route::patch('resources/tickets/?{id}?/{status}/update', [TicketsController::class, 'update'])->name('tickets.update')->where(['id' => 'num', 'status' => 'num']);
-
-        Route::post('resources/tickets/create', [TicketsController::class, 'create'])->name('tickets.create')->protected(Roles::ROLE[1], Roles::ROLE[2]);
+        Route::post('resources/tickets/create', [TicketsController::class, 'create'])->name('tickets.create')->protected(Roles::ROLE[0]);
         Route::post('resources/tickets/messages/create', [TicketsController::class, 'createMessage'])->name('tickets.messages.create');
         
         Route::delete('account/messages/?{id}?/delete', [MessagesController::class, 'delete'])->name('messages.delete')->where(['id' => 'num']);
@@ -82,22 +86,10 @@ Route::groupPrefix('admin', function () {
         Route::delete('account/activities/?{id}?/delete', [ActivitiesController::class, 'delete'])->name('activities.delete')->where(['id' => 'num']);
         Route::post('account/activities/export', [ActivitiesController::class, 'create'])->name('activities.export');
 
-        Route::groupProtected([Roles::ROLE[0], Roles::ROLE[1]], function () {
-            Route::delete('resources/users/?{id}?/delete', [UsersController::class, 'delete'])->name('users.delete')->where(['id' => 'num']);
-            Route::patch('resources/users/?{id}?/update', [UsersController::class, 'update'])->name('users.update')->where(['id' => 'num']);
-            Route::post('resources/users/create', [UsersController::class, 'create'])->name('users.create');
-            Route::post('resources/users/export', [UsersController::class, 'export'])->name('users.export');
-        });
-
-        Route::groupProtected([Roles::ROLE[0]], function () {
-            Route::delete('resources/invoices/?{id}?/delete', [InvoicesController::class, 'delete'])->name('invoices.delete')->where(['id' => 'num']);
-            Route::patch('resources/invoices/?{id}?/update', [InvoicesController::class, 'update'])->name('invoices.update')->where(['id' => 'num']);
-            Route::post('resources/invoices/create', [InvoicesController::class, 'create'])->name('invoices.create');
-        });
-
-        Route::groupProtected([Roles::ROLE[0], Roles::ROLE[1]], function () {
-            Route::post('resources/invoices/export', [InvoicesController::class, 'export'])->name('invoices.export');
-            Route::post('resources/invoices/{id}/download', [InvoicesController::class, 'download'])->name('invoices.download')->where(['id' => 'num']);
-        });
+        Route::delete('resources/invoices/?{id}?/delete', [InvoicesController::class, 'delete'])->name('invoices.delete')->where(['id' => 'num'])->protected(Roles::ROLE[0]);
+        Route::patch('resources/invoices/?{id}?/update', [InvoicesController::class, 'update'])->name('invoices.update')->where(['id' => 'num'])->protected(Roles::ROLE[0]);
+        Route::post('resources/invoices/create', [InvoicesController::class, 'create'])->name('invoices.create')->protected(Roles::ROLE[0]);
+        Route::post('resources/invoices/export', [InvoicesController::class, 'export'])->name('invoices.export')->protected(Roles::ROLE[2]);
+        Route::post('resources/invoices/{id}/download', [InvoicesController::class, 'download'])->name('invoices.download')->where(['id' => 'num'])->protected(Roles::ROLE[2]);
     });
 })->register();
